third_party/WebKit/LayoutTests/platform/win/virtual/service-worker-navigation-preload/http/tests/serviceworker/fetch-canvas-tainting-expected.txt

CONSOLE WARNING: line 14: Subresource requests whose URLs contain embedded credentials (e.g. `https://user:pass@host/`) are deprecated, and will be blocked in M59, around June 2017. See https://www.chromestatus.com/feature/5669008342777856 for more details.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&ignore' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&ignore' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&Auth&ignore' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access. The response had HTTP status code 401.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&ACAOrigin=http://127.0.0.1:8000&ignore' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&Auth&ACAOrigin=http://127.0.0.1:8000&ignore' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: The value of the 'Access-Control-Allow-Credentials' header in the response is '' which must be 'true' when the request's credentials mode is 'include'. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
CONSOLE ERROR: Access to Image at 'http://localhost:8000/serviceworker/resources/fetch-access-control.php?PNGIMAGE&Auth&ignore' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource. Origin 'http://127.0.0.1:8000' is therefore not allowed access.
This is a testharness.js-based test.
PASS Verify canvas tainting of fetched image in a Service Worker 
Harness: the test ran to completion.
